requirements.txt
# 기본 데이터 처리
pandas==2.1.4
numpy==1.26.2

# 주식 데이터 수집 (무료)
yfinance==0.2.33
pykrx==1.0.46
ccxt==4.2.25

# 기술적 분석 지표
pandas-ta==0.3.14b
ta==0.11.0

# 머신러닝
scikit-learn==1.3.2
scipy==1.11.4

# 웹 대시보드
streamlit==1.29.0
plotly==5.18.0

# 데이터 시각화
matplotlib==3.8.2

# API 요청
requests==2.31.0

# 날짜/시간 처리
pytz==2023.3.post1

# ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
# 🆕 5단계 추가 패키지 (2025-11-13)
# ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

# 웹 스크래핑 (SEC 데이터 수집용)
beautifulsoup4==4.12.2
lxml==4.9.3

# 설정 파일 관리
pyyaml==6.0.1
python-dotenv==1.0.0

# 유틸리티
tqdm==4.66.1
python-dateutil==2.8.2

# 알림 (선택 - 나중에 사용)
# schedule==1.2.0
# python-telegram-bot==20.7